% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtReg.R
\name{formula.gtReg}
\alias{formula.gtReg}
\title{Extract the model formula from a fitted group testing model}
\usage{
\method{formula}{gtReg}(x, ...)
}
\arguments{
\item{x}{An object of class "gtReg", created by \code{\link{gtReg}},
from which the model formula is to be extracted.}

\item{...}{not currently used.}
}
\value{
Model formula extracted from the object \kbd{object}.
}
\description{
Extract the model formula from objects of class "gtReg"
returned by \code{\link{gtReg}}.
}
\examples{
data(hivsurv)
fit1 <- gtReg(formula = groupres ~ AGE * EDUC.,
              data = hivsurv, groupn = gnum,
              linkf = "probit")
formula(x = fit1)
}
\author{
Brianna D. Hitt
}
