% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designFunctions.R
\name{print.designPower}
\alias{print.designPower}
\title{Print method for objects of class "designPower"}
\usage{
\method{print}{designPower}(x, ...)
}
\arguments{
\item{x}{an object of class "designPower" created by 
\code{\link{designPower}}.}

\item{...}{additional arguments to be passed to \code{print}. 
Currently only \code{digits} to be passed to \code{signif} for 
appropriate rounding.}
}
\value{
A print out detailing whether or not power was reached in 
the range of values (\kbd{n} or \kbd{s}) provided, the maximal 
power reached in the range of values, the alternative hypothesis, 
and the assumed true proportion.
}
\description{
Print method for objects of class "designPower" 
created by \code{\link{designPower}}.
}
\author{
This function was originally written as \code{print.bgtDesign} 
by Frank Schaarschmidt for the \code{binGroup} package. Minor 
modifications were made for inclusion in the \code{binGroup2} package.
}
