% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtReg.R
\name{print.gtReg}
\alias{print.gtReg}
\title{Print method for \kbd{gtReg}}
\usage{
\method{print}{gtReg}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{An object of class "gtReg" created by \code{\link{gtReg}}.}

\item{digits}{digits for rounding.}

\item{...}{currently not used.}
}
\value{
A print out of the function call, coefficients, and the null and
residual deviance and degrees of freedom.
}
\description{
Print method for objects obtained by \code{\link{gtReg}}.
}
\author{
This function was originally written by Boan Zhang as the
\kbd{print.gt} function for the \code{binGroup} package. Minor modifications
were made for inclusion in the \code{binGroup2} package.
}
