% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtTest.R
\name{print.gtTest}
\alias{print.gtTest}
\title{Print method for objects of class "gtTest"}
\usage{
\method{print}{gtTest}(x, ...)
}
\arguments{
\item{x}{An object of class "gtTest" (\code{\link{gtTest}}).}

\item{...}{Additional arguments to be passed to \code{print}. 
Currently only \code{digits} to be passed to \code{signif} for 
appropriate rounding.}
}
\value{
A print out of the p-value and point estimate resulting 
from \code{\link{gtTest}}.
}
\description{
Print method for objects of class "gtTest" created 
by the \code{\link{gtTest}} function.
}
\author{
This function was originally written as \code{print.bgtTest} by 
Brad Biggerstaff for the \code{binGroup} 
package. Minor modifications were made for inclusion of the function in 
the \code{binGroup2} package.
}
