% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halving.R
\name{print.halving}
\alias{print.halving}
\title{Print method for objects of class "halving"}
\usage{
\method{print}{halving}(x, ...)
}
\arguments{
\item{x}{An object of class "halving" (\code{\link{halving}}).}

\item{...}{Additional arguments to be passed to \code{print}.
Currently only \code{digits} to be passed to \code{signif} for
appropriate rounding.}
}
\value{
A print out of the PMF, expected testing expenditure and variance
of testing expenditure resulting from \code{\link{halving}}.
}
\description{
Print method for objects of class "halving" created
by the \code{\link{halving}} function.
}
\author{
Brianna D. Hitt
}
