% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opCharMethodFunctions.R
\name{print.opChar}
\alias{print.opChar}
\title{Print method for operating characteristics results}
\usage{
\method{print}{opChar}(x, ...)
}
\arguments{
\item{x}{an object of class \kbd{"opChar"}, providing the calculated
operating characteristics for a group testing algorithm.}

\item{...}{Additional arguments to be passed to \code{print} (e.g.,
\code{digits} to be passed to \code{round} for appropriate rounding).}
}
\value{
A print out of the algorithm, testing configuration, expected number
of tests, expected number of tests per individual, and accuracy measures
for individuals and for the overall algorithm.
}
\description{
Print method for objects of class \kbd{"opChar"} returned by
\code{\link{operatingCharacteristics1}} (\kbd{opChar1}) or
\code{\link{operatingCharacteristics2}} (\kbd{opChar2}).
}
\author{
Brianna D. Hitt
}
