% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OTCMethodFunctions.R
\name{print.OTC}
\alias{print.OTC}
\title{Print method for optimal testing configuration results}
\usage{
\method{print}{OTC}(x, ...)
}
\arguments{
\item{x}{an object of class \kbd{"OTC"}, providing the optimal testing
configuration results for a group testing algorithm.}

\item{...}{Additional arguments to be passed to \code{print} (e.g.,
\code{digits} to be passed to \code{round} for appropriate rounding).}
}
\value{
A print out of the algorithm, testing configuration, expected number
of tests, expected number of tests per individual, and accuracy measures
for individuals and for the overall algorithm.
}
\description{
Print method for objects of class \kbd{"OTC"} returned by
\code{\link{OTC1}} or \code{\link{OTC2}}.
}
\author{
Brianna D. Hitt
}
