% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtReg.R
\name{print.predict.gtReg}
\alias{print.predict.gtReg}
\title{Print method for \kbd{predict.gtReg}}
\usage{
\method{print}{predict.gtReg}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{An object of class "predict.gtReg" created by
\code{\link{predict.gtReg}}.}

\item{digits}{digits for rounding.}

\item{...}{not currently used.}
}
\value{
A matrix of predictions with rows corresponding to each observation
in \kbd{newdata} (if provided) or each observation in the data set used
for the fit. The columns correspond to the predictions (\kbd{fit}), the
estimated standard errors (\kbd{se.fit}), the lower bound of the confidence
interval (\kbd{lower}), and the upper bound of the confidence interval
(\kbd{upper}).
If \kbd{conf.level} is not specified, the \kbd{lower} and \kbd{upper}
columns will not be included. If \kbd{se = FALSE}, the \kbd{se.fit} column
will not be included.
}
\description{
Print method for objects obtained by
\code{\link{predict.gtReg}}.
}
\author{
Brianna Hitt
}
