% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propCI.R
\name{print.propCI}
\alias{print.propCI}
\title{Print method for objects of class "propCI"}
\usage{
\method{print}{propCI}(x, ...)
}
\arguments{
\item{x}{An object of class "propCI" (\code{\link{propCI}}).}

\item{...}{Additional arguments to be passed to \code{print}.}
}
\value{
A print out of the point estimate and confidence interval 
found with \code{\link{propCI}}.
}
\description{
Print method for objects of class "propCI" 
created by the \code{\link{propCI}} function.
}
\author{
This function is a combination of \code{print.poolbindiff} and 
\code{print.bgt}, written by Brad Biggerstaff for the \code{binGroup} 
package. Minor modifications were made for inclusion of the function in 
the \code{binGroup2} package.
}
