% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PooledBinomialFunctions.r
\name{print.propDiffCI}
\alias{print.propDiffCI}
\title{Print method for objects of class "propDiffCI"}
\usage{
\method{print}{propDiffCI}(x, ...)
}
\arguments{
\item{x}{An object of class "propDiffCI" (\code{\link{propDiffCI}}).}

\item{...}{Additional arguments to be passed to \code{print}.}
}
\value{
A print out of the point estimate and confidence interval 
found with \code{\link{propDiffCI}}.
}
\description{
Print method for objects of class "propDiffCI" 
created by the \code{\link{propDiffCI}} function.
}
\author{
This function was originally written as \code{print.poolbindiff}
by Brad Biggerstaff for the \code{binGroup} package. Minor 
modifications were made for inclusion of the function in the 
\code{binGroup2} package.
}
