% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SterrettFunctions.R
\name{print.Sterrett}
\alias{print.Sterrett}
\title{Print method for objects of class "Sterrett"}
\usage{
\method{print}{Sterrett}(x, ...)
}
\arguments{
\item{x}{An object of class "Sterrett" (\code{\link{Sterrett}}).}

\item{...}{Additional arguments to be passed to \code{print}.
Currently only \code{digits} to be passed to \code{signif} for
appropriate rounding.}
}
\value{
A print out of the PMF, expected testing expenditure and variance
of testing expenditure resulting from \code{\link{Sterrett}}.
}
\description{
Print method for objects of class "Sterrett" created
by the \code{\link{Sterrett}} function.
}
\author{
Brianna D. Hitt
}
