% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThresholdDorfman.R
\name{print.TOD}
\alias{print.TOD}
\title{Print method for \kbd{TOD}}
\usage{
\method{print}{TOD}(x, ...)
}
\arguments{
\item{x}{An object of class "TOD" created by
\code{\link{TOD}}.}

\item{...}{not currently used.}
}
\value{
A print out of configuration and operating characteristics
found with \code{\link{TOD}}.
}
\description{
Print method for objects obtained by
\code{\link{TOD}}.
}
\author{
Chris Bilder
}
